blathesis.cls: add print option
This commit adjusts the binding correction and margin depending on the
value of the print class option so that the effective textwidth with and
without the print option is identical. That way print and online
version will have the exact same number of pages and identical page
breaks. This allows us to have symmetric left/right margins for the
online version.

Signed-off-by: Niklas Koep <niklas.koep@gmail.com>